sistema_distribuido/requirements.txt
pyzmq
python-dateutil
pytest
filelock
